+++ src/lib/index.ts
@@ -2,6 +2,7 @@
 
 export { default as Stage } from './components/Stage.svelte'
 export { default as CodeEditor } from './components/CodeEditor.svelte'
+export { default as CodeLines } from './components/CodeLines.svelte'
 export { default as MarkdownIt } from './components/MarkdownIt.svelte'
 export { default as PresentationViewer } from './components/PresentationViewer.svelte'
 export { default as ViewerOverlay } from './components/ViewerOverlay.svelte'