+++ src/routes/magic-move/_slides/extract-learning.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-   import { MarkdownIt, slideLock } from '$lib/index.js'
+   import { MarkdownIt, slideLock, CodeEditor, typewriter, CodeLines } from '$lib/index.js'
    import content1 from './content/extract-learning-1.md?raw'
    import content2 from './content/extract-learning-2.md?raw'
    import content3 from './content/extract-learning-3.md?raw'
@@ -9,6 +9,18 @@
    let complete = $state(false)
    const contents = [ content1, content2, content3, content4 ]
    const content = $derived(contents[phase])
+
+   let code = $state('')
+   let code1 = $state('Hello there')
+   let code2 = $state('General Kenobi')
+
+   const outputs = {
+      ["children()"]: "error: Cannot read properties of undefined (reading 'before')",
+      ["children({ before: null })"]: "error: anchor.before is not a function",
+      ["children({ before: (...args) => console.dir(args) })"]: "logs either Nodes or a DocumentFragment",
+   }
+
+   const output = $derived(code in outputs ? outputs[code as keyof typeof outputs] : '')
 
    function prev() {
       if (phase === 0) return true
@@ -20,7 +32,7 @@
       phase = Math.min(3, phase+1)
    }
 </script>
-
+<!--
 <div id="markdown" class="markdown" {@attach slideLock(complete, prev, next)}>
    {#if phase === 0}
       <MarkdownIt content={content1} />
@@ -31,7 +43,13 @@
    {:else}
       <MarkdownIt content={content4} />
    {/if}
+</div> -->
+
+<div id="editor">
+   <CodeLines numLines={3} />
 </div>
+
+<div id="console" {@attach typewriter(output, 40)}></div>
 
 <!-- <div id="note">
    {#if phase === 0}
@@ -43,20 +61,56 @@
    <p>Ordinarily you render a snippet with render children(). So what if I just call that in my script tag?</p>
    <p>Use Svelte Playground iframe - show process until Node or DocumentFragment.</p>
    <p>render, in script tag, passing object with before, passing object with before function.</p>
+   <p>Create an OL of textareas... maybe can use CodeInput for each line? Can make some static then and hover.</p>
+   <p>Create a console output object to embed color info.</p>
 </template>
 
 <style>
-   #markdown {
+   /*#markdown {
       place-self: center;
       font-size: 2rem;
       text-align: left;
       width: 60cqw;
+   }*/
+
+   #editor {
+      place-self: center;
+      transform: translateY(-5cqh);
+      border: 1px solid grey;
+      /*padding: 1em;*/
+      color: white;
+      font-size: 2rem;
    }
 
-   #note {
+   #console {
+      place-self: center;
+      transform: translateY(15cqh);
+      font-size: 2rem;
+      background-color: #111;
+      border: 1px solid grey;
+      border-radius: 1em;
+      width: 50ch;
+      padding: 1em;
+      height: 3em;
+      color: lightcoral;
+      position: relative;
+   }
+
+   #console:before {
+      content: "console";
+      position: absolute;
+      color: white;
+      font-size: 0.5em;
+      top: 1em;
+      left: 1em;
+      color: grey;
+      font-family: monospace;
+   }
+
+   /*#note {
       color: orange;
       place-self: center;
       transform: translateY(10cqh);
       font-size: 2rem;
-   }
+   }*/
 </style>